improvement(format-phone): improvement after code-review
