+++ packages/utils/src/format-phone/index.ts
@@ -1,1 +1,4 @@
-export * from './util';
+import { formatPhoneNumber } from './format';
+import { getRawPhoneNumber } from './get-raw'
+
+export const phoneNumber = { format: formatPhoneNumber, getRaw: getRawPhoneNumber };
