+++ packages/utils/src/index.ts
@@ -5,3 +5,4 @@ export * from './get-currency-symbol';
 export * from './crop-account-number';
 export * from './pluralize';
 export * from './get-countries';
+export * from './format-phone';
